Revert "Refactor to avoid import loops"
This reverts commit aba68ba48a69480a915155879de1688647df0382.
