+++ app/app.go
@@ -10,7 +10,6 @@ import (
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/internal"
 	"fyne.io/fyne/v2/internal/app"
-	"fyne.io/fyne/v2/internal/build"
 	intRepo "fyne.io/fyne/v2/internal/repository"
 	"fyne.io/fyne/v2/storage"
 	"fyne.io/fyne/v2/storage/repository"
@@ -128,7 +127,7 @@ func New() fyne.App {
 	if meta.ID == "" {
 		checkLocalMetadata() // if no ID passed, check if it was in toml
 		if meta.ID == "" {
-			build.LogHint("Applications should be created with a unique ID using app.NewWithID()")
+			internal.LogHint("Applications should be created with a unique ID using app.NewWithID()")
 		}
 	}
 	return NewWithID(meta.ID)
