+++ app/app_mobile_and_test.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"testing"
 
-	"fyne.io/fyne/v2/internal/config"
+	"fyne.io/fyne/v2/internal/app"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -15,7 +15,7 @@ func Test_RootConfigDir(t *testing.T) {
 	oldEnv := os.Getenv("FILESPATH")
 	os.Setenv("FILESPATH", "/tmp")
 
-	assert.Equal(t, "/tmp", config.RootConfigDir())
+	assert.Equal(t, "/tmp", app.RootConfigDir())
 	os.Setenv("FILESPATH", oldEnv)
 }
