+++ app/preferences_android.go
@@ -5,7 +5,7 @@ package app
 import (
 	"path/filepath"
 
-	"fyne.io/fyne/v2/internal/config"
+	"fyne.io/fyne/v2/internal/app"
 )
 
 // storagePath returns the location of the settings storage
@@ -16,7 +16,7 @@ func (p *preferences) storagePath() string {
 
 // storageRoot returns the location of the app storage
 func (a *fyneApp) storageRoot() string {
-	return config.RootConfigDir() // we are in a sandbox, so no app ID added to this path
+	return app.RootConfigDir() // we are in a sandbox, so no app ID added to this path
 }
 
 func (p *preferences) watch() {
