+++ app/preferences_other.go
@@ -5,7 +5,7 @@ package app
 import (
 	"path/filepath"
 
-	"fyne.io/fyne/v2/internal/config"
+	"fyne.io/fyne/v2/internal/app"
 )
 
 // storagePath returns the location of the settings storage
@@ -15,7 +15,7 @@ func (p *preferences) storagePath() string {
 
 // storageRoot returns the location of the app storage
 func (a *fyneApp) storageRoot() string {
-	return filepath.Join(config.RootConfigDir(), a.UniqueID())
+	return filepath.Join(app.RootConfigDir(), a.UniqueID())
 }
 
 func (p *preferences) watch() {
