+++ app/settings.go
@@ -9,7 +9,6 @@ import (
 	"fyne.io/fyne/v2/internal/app"
 	"fyne.io/fyne/v2/internal/async"
 	"fyne.io/fyne/v2/internal/build"
-	"fyne.io/fyne/v2/internal/config"
 	"fyne.io/fyne/v2/theme"
 )
 
@@ -26,7 +25,7 @@ type SettingsSchema struct {
 
 // StoragePath returns the location of the settings storage
 func (sc *SettingsSchema) StoragePath() string {
-	return filepath.Join(config.RootConfigDir(), "settings.json")
+	return filepath.Join(app.RootConfigDir(), "settings.json")
 }
 
 // Declare conformity with Settings interface
@@ -130,7 +129,7 @@ func (s *settings) fileChanged() {
 }
 
 func (s *settings) loadSystemTheme() fyne.Theme {
-	path := filepath.Join(config.RootConfigDir(), "theme.json")
+	path := filepath.Join(app.RootConfigDir(), "theme.json")
 	data, err := fyne.LoadResourceFromPath(path)
 	if err != nil {
 		if !os.IsNotExist(err) {
