+++ container/multiplewindows_test.go
@@ -1,26 +1,25 @@
-package container_test
+package container
 
 import (
 	"testing"
 
 	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/container"
 	"fyne.io/fyne/v2/test"
 	"fyne.io/fyne/v2/widget"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestMultipleWindows_Add(t *testing.T) {
-	m := container.NewMultipleWindows()
+	m := NewMultipleWindows()
 	assert.Zero(t, len(m.Windows))
 
-	m.Add(container.NewInnerWindow("1", widget.NewLabel("Inside")))
+	m.Add(NewInnerWindow("1", widget.NewLabel("Inside")))
 	assert.Equal(t, 1, len(m.Windows))
 }
 
 func TestMultipleWindows_Drag(t *testing.T) {
-	w := container.NewInnerWindow("1", widget.NewLabel("Inside"))
-	m := container.NewMultipleWindows(w)
+	w := NewInnerWindow("1", widget.NewLabel("Inside"))
+	m := NewMultipleWindows(w)
 	_ = test.TempWidgetRenderer(t, m) // initialise display
 	assert.Equal(t, 1, len(m.Windows))
