--- container/tabs_internal_test.go
@@ -1,22 +1,0 @@
-package container
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	"fyne.io/fyne/v2/canvas"
-	"fyne.io/fyne/v2/internal/cache"
-	"fyne.io/fyne/v2/theme"
-)
-
-func TestTabButton_Icon_Change(t *testing.T) {
-	b := &tabButton{icon: theme.CancelIcon()}
-	r := cache.Renderer(b)
-	icon := r.Objects()[3].(*canvas.Image)
-	oldResource := icon.Resource
-
-	b.icon = theme.ConfirmIcon()
-	b.Refresh()
-	assert.NotEqual(t, oldResource, icon.Resource)
-}
