+++ container/tabs_test.go
@@ -1,25 +1,37 @@
-package container_test
+package container
 
 import (
 	"testing"
 
-	"fyne.io/fyne/v2/container"
 	internalTest "fyne.io/fyne/v2/internal/test"
-	"fyne.io/fyne/v2/test"
 	"github.com/stretchr/testify/assert"
 
 	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/canvas"
+	"fyne.io/fyne/v2/internal/cache"
+	"fyne.io/fyne/v2/test"
 	"fyne.io/fyne/v2/theme"
 	"fyne.io/fyne/v2/widget"
 )
+
+func TestTabButton_Icon_Change(t *testing.T) {
+	b := &tabButton{icon: theme.CancelIcon()}
+	r := cache.Renderer(b)
+	icon := r.Objects()[3].(*canvas.Image)
+	oldResource := icon.Resource
+
+	b.icon = theme.ConfirmIcon()
+	b.Refresh()
+	assert.NotEqual(t, oldResource, icon.Resource)
+}
 
 func TestTab_ThemeChange(t *testing.T) {
 	a := test.NewTempApp(t)
 	a.Settings().SetTheme(internalTest.LightTheme(theme.DefaultTheme()))
 
-	tabs := container.NewAppTabs(
-		container.NewTabItem("a", widget.NewLabel("a")),
-		container.NewTabItem("b", widget.NewLabel("b")))
+	tabs := NewAppTabs(
+		NewTabItem("a", widget.NewLabel("a")),
+		NewTabItem("b", widget.NewLabel("b")))
 	w := test.NewTempWindow(t, tabs)
 	w.Resize(fyne.NewSize(180, 120))
