+++ internal/app/config.go
@@ -1,4 +1,4 @@
-package config
+package app
 
 func RootConfigDir() string {
 	return rootConfigDir()
