+++ internal/app/config_desktop_darwin.go
@@ -1,6 +1,6 @@
 //go:build !ci && !ios && !wasm && !test_web_driver && !mobile
 
-package config
+package app
 
 import (
 	"os"
