+++ internal/app/config_wasm.go
@@ -1,6 +1,6 @@
 //go:build !ci && (!android || !ios || !mobile) && (wasm || test_web_driver)
 
-package config
+package app
 
 func rootConfigDir() string {
 	return "/data/"
