+++ internal/app/config_xdg.go
@@ -1,6 +1,6 @@
 //go:build !ci && !wasm && !test_web_driver && !android && !ios && !mobile && (linux || openbsd || freebsd || netbsd)
 
-package config
+package app
 
 import (
 	"os"
